Organise condition code
