+++ src/parser/condition/condition-val.js
@@ -1,0 +1,6 @@
+const P = require('parsimmon')
+const { ConditionExpr } = require('./condition-expr')
+
+module.exports = {
+  ConditionVal: null
+}
