+++ src/parser/condition/function.js
@@ -1,0 +1,7 @@
+const P = require('parsimmon')
+const { Str, _ } = require('../shared')
+const { RelName } = require('../relationship')
+
+function AbstractFunction(name, arguments) {
+  return
+}
